Restrict tables to available output width
Restrict tables, that are printed to stdout, to the available output
width. Currently that is `usize::MAX` for anything printed directly from
wash cli and the output pane width inside the wash up REPL.

Restricting tables means columns shrink so all columns fit in one line.
If a value inside a column is too long, it breaks into multiple lines
inside the column only.

Example before (| is the terminal width):

    | Actor ID                                            Image Refere |
    | nce                                                              |
    | ABCDEFGHIJABCDEFGHIJABCDEFGHIJABCDEFGHIJABCDEFGHIJ  wasmcloud.az |
    | urecr.io/kvcounter:0.2.0                                         |

Example after (| is the terminal width):

    | Actor ID                           Image Reference               |
    | ABCDEFGHIJABCDEFGHIJABCDEFGHIJABC  wasmcloud.azurecr.io/kvcounte |
    | DEFGHIJABCDEFGHIJ                  r:0.2.0                       |

In a later change we could change the behavior for overflowing column
values. For example instead of wrapping we could show the first few
characters followed by an ellipsis.